requirements.txt
streamlit
scikit-learn
pandas
numpy
plotly
sqlalchemy
python-dotenv
pymysql
